Drop unnecessary usize cast in Rc4::new
`counter` is inferred to `usize` from its use as an array index into
`state`, so the explicit `as usize` cast on the same line is
redundant and clippy's `unnecessary_cast` lint flags it.

+++ src/v1/streamcipher/crypto/rc4.rs
@@ -50,7 +50,7 @@ impl Rc4 {
         let mut index2 = 0u8;
         for counter in 0..256 {
             index2 = key[index1 as usize].wrapping_add(state[counter]).wrapping_add(index2);
-            state.swap(counter as usize, index2 as usize);
+            state.swap(counter, index2 as usize);
             index1 = index1.wrapping_add(1) % key_len;
         }
